refactor: dto 클래스 세분화
