+++ src/main/java/com/example/outsourcing/domain/order/dto/request/OrderRequestDto.java
@@ -4,6 +4,8 @@ import com.example.outsourcing.domain.order.entity.Orders;
 import com.example.outsourcing.domain.order.enums.OrderStatus;
 import com.example.outsourcing.domain.store.entity.Store;
 import com.example.outsourcing.domain.user.entity.User;
+import jakarta.persistence.Column;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
@@ -12,6 +14,7 @@ import java.util.List;
 @Getter
 public class OrderRequestDto {
 
+    @NotNull
     private List<OrderItemRequestDto> orderItems;
     private Integer usedPoint;
